CDev extract from drivers Device class hierarchy
 - leave CDev shim for now until all usage is updated
