+++ src/lib/drivers/device/CMakeLists.txt
@@ -33,9 +33,6 @@
 
 set(SRCS_PLATFORM)
 if (${OS} STREQUAL "nuttx")
-	list(APPEND SRCS_PLATFORM
-		nuttx/cdev_platform.cpp
-	)
 	if ("${CONFIG_I2C}" STREQUAL "y")
 		list(APPEND SRCS_PLATFORM nuttx/I2C.cpp)
 	endif()
@@ -46,8 +43,6 @@ if (${OS} STREQUAL "nuttx")
 else()
 	list(APPEND SRCS_PLATFORM
 		posix/I2C.cpp
-		posix/cdev_platform.cpp
-		posix/vfile.cpp
 	)
 endif()
 
@@ -61,3 +56,5 @@ px4_add_library(drivers__device
 if (${OS} STREQUAL "nuttx")
 	target_link_libraries(drivers__device PRIVATE nuttx_arch)
 endif()
+
+target_link_libraries(drivers__device PRIVATE cdev)
