+++ src/lib/drivers/device/Device.hpp
@@ -137,21 +137,23 @@ public:
 	 *
 	 * @return The bus ID
 	 */
-	uint8_t get_device_bus() { return _device_id.devid_s.bus; }
+	uint8_t get_device_bus() const { return _device_id.devid_s.bus; }
+
+	uint32_t get_device_id() const { return _device_id.devid; }
 
 	/**
 	 * Return the bus type the device is connected to.
 	 *
 	 * @return The bus type
 	 */
-	DeviceBusType get_device_bus_type() { return _device_id.devid_s.bus_type; }
+	DeviceBusType get_device_bus_type() const { return _device_id.devid_s.bus_type; }
 
 	/**
 	 * Return the bus address of the device.
 	 *
 	 * @return The bus address
 	 */
-	uint8_t get_device_address() { return _device_id.devid_s.address; }
+	uint8_t get_device_address() const { return _device_id.devid_s.address; }
 
 	void set_device_address(int address) { _device_id.devid_s.address = address; }
